requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
cx-Oracle==8.3.0
execnet==1.9.0
Faker==13.15.0
future==0.18.2
glob2==0.7
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
loguru==0.6.0
Mako==1.2.0
MarkupSafe==2.1.1
MouseInfo==0.1.3
packaging==21.3
parse==1.19.0
parse-type==0.6.0
playwright==1.22.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyAutoGUI==0.9.53
pyee==8.1.0
PyGetWindow==0.0.9
PyHive==0.6.5
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==7.1.2
pytest-base-url==2.0.0
pytest-bdd==5.0.0
pytest-forked==1.4.0
pytest-parallel==0.0.10
pytest-playwright==0.3.0
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-slugify==6.1.2
pytweening==1.0.4
PyYAML==6.0
requests==2.27.1
six==1.16.0
tblib==1.7.0
text-unidecode==1.3
thrift==0.16.0
tomli==2.0.1
urllib3==1.26.9
websockets==10.1
win32-setctime==1.1.0
